Fixes to remove compiler warning 4244 when compiling with vs2010

+++ src/ptlib/common/collect.cxx
@@ -51,7 +51,7 @@ PDEFINE_POOL_ALLOCATOR(PHashTableElement)
 void PCollection::PrintOn(ostream &strm) const
 {
   char separator = strm.fill();
-  int width = strm.width();
+  int width = (int)strm.width();
   for (PINDEX  i = 0; i < GetSize(); i++) {
     if (i > 0 && separator != ' ')
       strm << separator;
