+++ src/ptlib/common/ptime.cxx
@@ -75,7 +75,7 @@ PObject::Comparison PTimeInterval::Compare(const PObject & obj) const
 
 void PTimeInterval::PrintOn(ostream & stream) const
 {
-  int precision = stream.precision();
+  int precision = (int)stream.precision();
 
   Formats fmt = NormalFormat;
   if ((stream.flags()&ios::scientific) != 0)
@@ -85,7 +85,7 @@ void PTimeInterval::PrintOn(ostream & stream) const
     precision = -precision;
   }
 
-  stream << AsString(precision, fmt, stream.width());
+  stream << AsString(precision, fmt, (int)stream.width());
 }
